Cast one of the operands of this subtraction operation to a 'long'

+++ src/main/java/es/codeurjc/model/User.java
@@ -190,7 +190,7 @@ public class User {
     
     public long getYearsWithBank() {
         if (registrationDate == null) return 0;
-        return LocalDate.now().getYear() - registrationDate.getYear();
+        return (long)LocalDate.now().getYear() - registrationDate.getYear();
     }
     
     public boolean hasMultipleProducts() {
